src/requirements.txt
numpy
openai
openai-whisper
simpleaudio
sounddevice
soundfile
python-dotenv
pydub
pynput
pvporcupine
piper-tts
requests
